pyproject.toml
[project]
name = "second-project-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.1",
    "gitingest>=0.3.1",
    "httpx>=0.28.1",
    "langchain-core>=1.3.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "pygithub>=2.9.1",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
